placed the global context in a separate file to solve the cyclic dependency
